It was a mistake in the device print, trying now with doubles

+++ sym_test.cu
@@ -116,9 +116,9 @@ void open_files(string name_file, int pts, PointW3D *datos, float &size_box){
 }
 
 
-__global__ void simmetrization(float *s_XXX,float *XXX , int bn){
+__global__ void simmetrization(double *s_XXX,double *XXX , int bn){
     int i = blockIdx.x * blockDim.x + threadIdx.x;
-    float v;
+    double v;
     printf("%f \n", XXX[i]);
     for (int j=i+1; j<bn; j++){
         for (int k=j+1; k<bn; k++){
@@ -141,10 +141,10 @@ int main(int argc, char **argv){
     clock_t start_timmer, stop_timmer;
     double time_spent;
 
-    float *DDD, *d_DDD, *sd_DDD;
-    DDD = new float[bn*bn*bn];
-    cucheck(cudaMalloc(&d_DDD, bn*bn*bn*sizeof(float)));
-    cucheck(cudaMalloc(&sd_DDD, bn*bn*bn*sizeof(float)));
+    double *DDD, *d_DDD, *sd_DDD;
+    DDD = new double[bn*bn*bn];
+    cucheck(cudaMalloc(&d_DDD, bn*bn*bn*sizeof(double)));
+    cucheck(cudaMalloc(&sd_DDD, bn*bn*bn*sizeof(double)));
 
     PointW3D *dataD;
     dataD = new PointW3D[np];
@@ -160,7 +160,7 @@ int main(int argc, char **argv){
 
     cout << "Entering the device code \n" << endl;
 
-    cucheck(cudaMemcpy(d_DDD, DDD, bn*bn*bn*sizeof(float), cudaMemcpyHostToDevice));
+    cucheck(cudaMemcpy(d_DDD, DDD, bn*bn*bn*sizeof(double), cudaMemcpyHostToDevice));
 
     start_timmer = clock();
 
@@ -171,7 +171,7 @@ int main(int argc, char **argv){
     time_spent = (double)(stop_timmer - start_timmer) / CLOCKS_PER_SEC;
     printf("\nSpent time = %.4f seg.\n", time_spent );
 
-    cucheck(cudaMemcpy(DDD, sd_DDD, bn*bn*bn*sizeof(float), cudaMemcpyDeviceToHost));
+    cucheck(cudaMemcpy(DDD, sd_DDD, bn*bn*bn*sizeof(double), cudaMemcpyDeviceToHost));
     
     save_histogram("DDD.dat", bn, DDD);
